requirements.txt
moviepy==1.0.3
ffmpeg-python
pyttsx3

whisperx
pyannote-audio
torch
torchaudio

google-generativeai 
python-dotenv

flask
flask_socketio